delete 검증 추가

+++ src/main/java/org/deneb/springbootreact/student/StudentService.java
@@ -2,9 +2,11 @@ package org.deneb.springbootreact.student;
 
 import lombok.AllArgsConstructor;
 import org.deneb.springbootreact.student.exception.BadRequestException;
+import org.deneb.springbootreact.student.exception.StudentNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -26,6 +28,12 @@ public class StudentService {
   }
 
   public void delete(Long studentId) {
+    boolean isExist = studentRepository.findById(studentId).isPresent();
+    if (!isExist) {
+      throw new StudentNotFoundException(
+        String.format("Student not found with id : %s", studentId));
+    }
+
     studentRepository.deleteById(studentId);
   }
